fix home page drawing
