+++ package/flaskapp/dash_simtool/app/dashboard_callbacks.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+
+from dash.dependencies import Output, Input
+
+import package.flaskapp.dash_simtool.app.dashboard_styling as styling
+
+
+def add_legend_button(dash_app):
+    @dash_app.callback(
+        [
+            Output("legend", "style"),
+        ],
+
+        [Input("legend_button", "n_clicks")],
+        [
+
+        ]
+    )
+    def toggle_legend(n):
+        if(n!=None):
+            if(n%2 == 1):
+                return [styling.LEGEND]
+            else:
+                return [styling.LEGEND_HIDDEN]
+        else:
+            return [styling.LEGEND_HIDDEN]
+
+    return dash_app
