+++ package/flaskapp/dash_simtool/app/dashboard_components.py
@@ -333,22 +333,22 @@ def navbar_controls(url_page):
         [
             dbc.Nav(
                 [
-                    # dbc.NavLink("Home",
-                    #             href=dash_app.URL_SLDS if cf.demo else dash_app.URL_HOME,
-                    #             active=True if dash_app.URL_HOME == url_page else False,
-                    #             external_link=True),
+                    dbc.NavLink("Home",
+                                href=dash_app.URL_SLDS if cf.demo else dash_app.URL_HOME,
+                                active=True if dash_app.URL_HOME == url_page else False,
+                                external_link=True),
                     dbc.NavLink("SLDs",
                                 href=dash_app.URL_SLDS,
                                 active=True if dash_app.URL_SLDS == url_page else False,
                                 external_link=True),
-                    # dbc.NavLink("Scripts",
-                    #             href=dash_app.URL_SLDS if cf.demo else dash_app.URL_SCRIPTS,
-                    #             active=True if dash_app.URL_SCRIPTS == url_page else False,
-                    #             external_link=True),
-                    # dbc.NavLink("About",
-                    #             href=dash_app.URL_SLDS if cf.demo else dash_app.URL_ABOUT,
-                    #             active=True if dash_app.URL_ABOUT == url_page else False,
-                    #             external_link=True),
+                    dbc.NavLink("Scripts",
+                                href=dash_app.URL_SLDS if cf.demo else dash_app.URL_SCRIPTS,
+                                active=True if dash_app.URL_SCRIPTS == url_page else False,
+                                external_link=True),
+                    dbc.NavLink("About",
+                                href=dash_app.URL_SLDS if cf.demo else dash_app.URL_ABOUT,
+                                active=True if dash_app.URL_ABOUT == url_page else False,
+                                external_link=True),
                     dbc.NavLink("Log out",
                                 href='/logout',
                                 external_link=True),
